docs: Add suggestion to use awful.spawn’s async support
This isn’t really an issue beyond initial cache generation, but it would
be a lot cleaner.  I’m uninterested in writing it unless someone asks
though, as it nesting callbacks for the generation will really hurt
readability for what is frankly simply supposed to be an example of how
to implement timed wallpaper support.